+++ src/main/java/com/flechazo/slashblade/util/RayTraceHelper.java
@@ -1,22 +1,21 @@
 package com.flechazo.slashblade.util;
 
 import net.minecraft.world.entity.Entity;
-import net.minecraft.world.phys.Vec3;
+import net.minecraft.world.level.ClipContext;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.phys.AABB;
+import net.minecraft.world.phys.EntityHitResult;
+import net.minecraft.world.phys.HitResult;
+import net.minecraft.world.phys.Vec3;
 
 import javax.annotation.Nullable;
 import java.util.Optional;
 import java.util.function.Predicate;
 
-import net.minecraft.world.level.ClipContext;
-import net.minecraft.world.phys.AABB;
-import net.minecraft.world.phys.EntityHitResult;
-import net.minecraft.world.phys.HitResult;
-
 public class RayTraceHelper {
 
     public static Optional<HitResult> rayTrace(Level worldIn, Entity entityIn, Vec3 start, Vec3 dir, double blockReach,
-            double entityReach, Predicate<Entity> selector) {
+                                               double entityReach, Predicate<Entity> selector) {
         Vec3 end = start.add(dir.scale(blockReach));
 
         HitResult raytraceresult = worldIn
@@ -40,18 +39,18 @@ public class RayTraceHelper {
 
     @Nullable
     public static EntityHitResult rayTrace(Level worldIn, Entity entityIn, Vec3 start, Vec3 end, AABB boundingBox,
-            Predicate<Entity> selector) {
+                                           Predicate<Entity> selector) {
         return rayTrace(worldIn, entityIn, start, end, boundingBox, selector, Double.MAX_VALUE);
     }
 
     @Nullable
     public static EntityHitResult rayTrace(Level worldIn, Entity entityIn, Vec3 start, Vec3 end, AABB boundingBox,
-            Predicate<Entity> selector, double limitDist) {
+                                           Predicate<Entity> selector, double limitDist) {
         double currentDist = limitDist;
         Entity resultEntity = null;
 
         for (Entity foundEntity : worldIn.getEntities(entityIn, boundingBox, selector)) {
-            AABB axisalignedbb = foundEntity.getBoundingBox().inflate((double) 0.5F);
+            AABB axisalignedbb = foundEntity.getBoundingBox().inflate(0.5F);
             Optional<Vec3> optional = axisalignedbb.clip(start, end);
             if (optional.isPresent()) {
                 double newDist = start.distanceToSqr(optional.get());
